+++ src/aeon2yw_novelyst.py
@@ -77,14 +77,14 @@ class Plugin():
         self._pluginMenu = tk.Menu(self._ui.mainMenu, tearoff=0)
         self._ui.mainMenu.add_cascade(label=APPLICATION, menu=self._pluginMenu)
         self._ui.mainMenu.entryconfig(APPLICATION, state='disabled')
-        self._pluginMenu.add_command(label=_('Information'), underline=0, command=self._info)
+        self._pluginMenu.add_command(label=_('Information'), command=self._info)
         self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Update the timeline'), underline=7, command=self._export_from_yw)
-        self._pluginMenu.add_command(label=_('Update the project'), underline=7, command=self._import_to_yw)
+        self._pluginMenu.add_command(label=_('Update the timeline'), command=self._export_from_yw)
+        self._pluginMenu.add_command(label=_('Update the project'), command=self._import_to_yw)
         self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Add or update moon phase data'), underline=14, command=self._add_moonphase)
+        self._pluginMenu.add_command(label=_('Add or update moon phase data'), command=self._add_moonphase)
         self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Edit the timeline'), underline=0, command=self._launch_application)
+        self._pluginMenu.add_command(label=_('Edit the timeline'), command=self._launch_application)
 
     def disable_menu(self):
         """Disable menu entries when no project is open."""