+++ src/make_pot.py
@@ -1,12 +1,12 @@
 """Generate a template file (pot) for message translation.
 
-For further information see https://github.com/peter88213/novelyst
+For further information see https://github.com/peter88213/aeon2yw
 Published under the MIT License (https://opensource.org/licenses/mit-license.php)
 """
 import os
 import sys
-import build_aeon2yw
-import build_aeon2yw_novelyst
+import build_aeon2yw as build_application
+import build_aeon2yw_novelyst as build_plugin
 sys.path.insert(0, f'{os.getcwd()}/../../PyWriter/src')
 import pgettext
 
@@ -16,9 +16,9 @@ POT_FILE = '../i18n/messages.pot'
 
 def make_pot(version='unknown'):
     # Generate a complete script.
-    build_aeon2yw.main()
+    build_application.main()
     # Generate a complete plugin.
-    build_aeon2yw_novelyst.main()
+    build_plugin.main()
 
     # Generate a pot file from the script.
     if os.path.isfile(POT_FILE):
@@ -28,8 +28,8 @@ def make_pot(version='unknown'):
         backedUp = False
     try:
         pot = pgettext.PotFile(POT_FILE, app=APP, appVersion=version)
-        pot.scan_file(build_aeon2yw.TARGET_FILE)
-        pot.scan_file(build_aeon2yw_novelyst.TARGET_FILE)
+        pot.scan_file(build_application.TARGET_FILE)
+        pot.scan_file(build_plugin.TARGET_FILE)
         print(f'Writing "{pot.filePath}"...\n')
         pot.write_pot()